requirements.txt
# Python dependencies for the Image Classification API
fastapi==0.95.2
uvicorn==0.23.2
tensorflow==2.13.0
pillow==10.0.0
python-multipart==0.0.6
numpy==1.24.3
typing-extensions==4.5.0
pydantic==1.10.8
jinja2==3.1.2

